+++ pkg/features/list.go
@@ -118,6 +118,9 @@ var (
 
 	// Adds the ability to generate on-demand vulnerability reports based on filter views
 	VulnerabilityOnDemandReports = registerFeature("Adds the ability to generate on-demand vulnerability reports based on filter views", "ROX_VULNERABILITY_ON_DEMAND_REPORTS")
+
+	// Adds the ability to customize the regex rules for identifying platform components
+	CustomizablePlatformComponents = registerFeature("Adds the ability to customize the regex rules for identifying platform components", "ROX_CUSTOMIZABLE_PLATFORM_COMPONENTS")
 )
 
 // The following feature flags are related to Scanner V4.
@@ -153,4 +156,7 @@ var (
 	//
 	// This must be set in Scanner V4 Matcher to have any effect.
 	ScannerV4RedHatCSAF = registerFeature("Scanner V4 will enrich its results with Red Hat CSAF data", "ROX_SCANNER_V4_RED_HAT_CSAF", enabled)
+
+	// ScannerV4RedHatLayers enables displaying vulnerabilities from Red Hat sources, only, for packages found in official Red Hat image layers.
+	ScannerV4RedHatLayers = registerFeature("Scanner V4 will only output vulnerabilities from Red Hat sources, only, for packages found in official Red Hat image layers", "ROX_SCANNER_V4_RED_HAT_LAYERS_RED_HAT_VULNS_ONLY", enabled)
 )